requirements.txt
python-decouple

django
djangorestframework
djangorestframework-simplejwt

psycopg2-binary

celery 
redis

datasets
beautifulsoup4

langchain 
langchain-community 
faiss-cpu 
sentence-transformers
langchain-google-genai